move getValidationModes to logic folder
