+++ src/useFieldArray.ts
@@ -3,6 +3,7 @@ import React from 'react';
 import focusFieldBy from './logic/focusFieldBy';
 import generateId from './logic/generateId';
 import getFocusFieldName from './logic/getFocusFieldName';
+import getValidationModes from './logic/getValidationModes';
 import isWatched from './logic/isWatched';
 import updateFieldArrayRootError from './logic/updateFieldArrayRootError';
 import validateField from './logic/validateField';
@@ -11,7 +12,6 @@ import cloneObject from './utils/cloneObject';
 import convertToArrayPayload from './utils/convertToArrayPayload';
 import fillEmptyArray from './utils/fillEmptyArray';
 import get from './utils/get';
-import getValidationModes from './utils/getValidationModes';
 import insertAt from './utils/insert';
 import isEmptyObject from './utils/isEmptyObject';
 import moveArrayAt from './utils/move';
